+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Changelog
 
 ## Version 2.5.0-beta2
-- Improved ServiceBus depedency and requests collection for (Microsoft.Azure.ServiceBus 2.1.0)
+- Improved ServiceBus dependency and requests collection for (Microsoft.Azure.ServiceBus 2.1.0)
 
 ## Version 2.5.0-beta1
 - Removed `net40` targets from all packages. Use the version 2.4 of SDK if your application is still compiled with the framework 4.0.